Unison.Sqlite.Connection

Connection management

data Connection

A non-thread safe connection to a SQLite database.

Constructors

Connection

Fields

name :: String

file :: FilePath

conn :: Connection

withConnection

Arguments

:: MonadUnliftIO m

=> String
Connection name, for debugging.

-> FilePath
Path to SQLite database file.

-> (Connection -> m a)

-> m a

Perform an action with a connection to a SQLite database.

Note: the connection is created with PRAGMA foreign_keys = ON automatically, to work around the fact that SQLite does not automatically enforce foreign key integrity, because it elected to maintain backwards compatibility with code that was written before the foreign key integrity feature was implemented.

Executing queries

Without results

execute :: HasCallStack => Connection -> Sql -> IO ()

executeStatements :: HasCallStack => Connection -> Text -> IO ()

Execute one or more semicolon-delimited statements.

This function does not support parameters, and is mostly useful for executing DDL and migrations.

With results

queryStreamRow :: (HasCallStack, FromRow a) => Connection -> Sql -> (IO (Maybe a) -> IO r) -> IO r

queryStreamCol :: forall a r. (HasCallStack, FromField a) => Connection -> Sql -> (IO (Maybe a) -> IO r) -> IO r

queryListRow :: forall a. (FromRow a, HasCallStack) => Connection -> Sql -> IO [a]

queryListCol :: forall a. (FromField a, HasCallStack) => Connection -> Sql -> IO [a]

queryMaybeRow :: (FromRow a, HasCallStack) => Connection -> Sql -> IO (Maybe a)

queryMaybeCol :: forall a. (FromField a, HasCallStack) => Connection -> Sql -> IO (Maybe a)

queryOneRow :: (FromRow a, HasCallStack) => Connection -> Sql -> IO a

queryOneCol :: forall a. (FromField a, HasCallStack) => Connection -> Sql -> IO a

With checks

queryListRowCheck :: (FromRow a, SqliteExceptionReason e, HasCallStack) => Connection -> Sql -> ([a] -> Either e r) -> IO r

queryListColCheck :: forall a e r. (FromField a, SqliteExceptionReason e, HasCallStack) => Connection -> Sql -> ([a] -> Either e r) -> IO r

queryMaybeRowCheck :: (FromRow a, SqliteExceptionReason e, HasCallStack) => Connection -> Sql -> (a -> Either e r) -> IO (Maybe r)

queryMaybeColCheck :: forall a e r. (FromField a, SqliteExceptionReason e, HasCallStack) => Connection -> Sql -> (a -> Either e r) -> IO (Maybe r)

queryOneRowCheck :: (FromRow a, SqliteExceptionReason e, HasCallStack) => Connection -> Sql -> (a -> Either e r) -> IO r

queryOneColCheck :: forall a e r. (FromField a, SqliteExceptionReason e, HasCallStack) => Connection -> Sql -> (a -> Either e r) -> IO r

Rows modified

rowsModified :: Connection -> IO Int

Vacuum (into)

vacuum :: Connection -> IO Bool

VACUUM, and return whether or not the vacuum succeeded. A vacuum fails if the connection has any open transactions.

vacuumInto :: Connection -> FilePath -> IO ()

VACUUM INTO

Low-level operations

Transaction

begin :: Connection -> IO ()

BEGIN

beginImmediate :: Connection -> IO ()

BEGIN IMMEDIATE

commit :: Connection -> IO ()

COMMIT

rollback :: Connection -> IO ()

ROLLBACK

Savepoint

withSavepoint :: MonadUnliftIO m => Connection -> Text -> (m () -> m a) -> m a

Perform an action within a named savepoint. The action is provided a rollback action.

withSavepointIO :: Connection -> Text -> (IO () -> IO a) -> IO a

savepoint :: Connection -> Text -> IO ()

SAVEPOINT

rollbackTo :: Connection -> Text -> IO ()

ROLLBACK TO

release :: Connection -> Text -> IO ()

RELEASE

Exceptions

newtype ExpectedAtMostOneRowException

A query was expected to return exactly one row, but it did not. The exception carries a string representation of the rows that were actually returned.

Constructors

ExpectedAtMostOneRowException

Fields

rows :: String

newtype ExpectedExactlyOneRowException

A query was expected to return exactly one row, but it did not. The exception carries a string representation of the rows that were actually returned.

Constructors

ExpectedExactlyOneRowException

Fields

rows :: String