Unison.Sqlite

Description

The Unison monorepo interface to SQLite.

This module provides a high(-er) level interface to SQLite than the sqlite-simple library, which it wraps. Code that interacts with SQLite in this monorepo should use this interface, rather than sqlite-simple or direct-sqlite directly.

Three variants of the main query interface are provided:

Unison.Sqlite.Connection provides an interface in IO, which takes the Connection argument as an explicit argument.

Unison.Sqlite.DB provides a type class interface, which moves the Connection to an implicit argument. This interface is also re-exported by this module, for convenient backwards compatibility with the existing queries.

Unison.Sqlite.Transaction provides a newer, yet-unused interface that executes queries in transactions, with automatic retries on SQLITE_BUSY due to concurrent writers.

Connection management

data Connection

A non-thread safe connection to a SQLite database.

withConnection

Arguments

:: MonadUnliftIO m

=> String
Connection name, for debugging.

-> FilePath
Path to SQLite database file.

-> (Connection -> m a)

-> m a

Perform an action with a connection to a SQLite database.

Note: the connection is created with PRAGMA foreign_keys = ON automatically, to work around the fact that SQLite does not automatically enforce foreign key integrity, because it elected to maintain backwards compatibility with code that was written before the foreign key integrity feature was implemented.

Type class query interface

type DB m = (MonadIO m, MonadReader Connection m)

runDB :: MonadIO m => Connection -> ReaderT Connection m a -> m a

Executing queries

newtype Sql

A SQL snippet.

Constructors

Sql Text

Without results

With parameters

execute :: (DB m, ToRow a) => Sql -> a -> m ()

executeMany :: (DB m, ToRow a) => Sql -> [a] -> m ()

Without parameters

execute_ :: DB m => Sql -> m ()

With results

Queries that return results have many different variants.

Every function name begins with the string query.

Row count. The caller may expect exactly one, zero or one, or zero or more rows, in which case the function name includes the string List, Maybe, or One, respectively. Example: queryListRow.

Row width. The caller may expect the returned rows may contain exactly one or more than one column, in which case the function name includes the string Col or Row, respectively. Example: queryOneCol.

Result checks. The caller may want to perform additional validation on the returned rows, in which case the function name includes the string Check. Example: queryMaybeColCheck.

Parameter count. The query may contain zero or one or more parameters. In the former case, the function name includes the string _. Example: queryListRow_.

All together, the full anatomy of a query function is:

query(List|Maybe|One)(Row|Col)[Check][_]

With parameters

queryListRow :: (DB m, FromRow a, ToRow b) => Sql -> b -> m [a]

queryListCol :: (DB m, FromField a, ToRow b) => Sql -> b -> m [a]

queryMaybeRow :: (DB m, FromRow a, ToRow b) => Sql -> b -> m (Maybe a)

queryMaybeCol :: (DB m, FromField a, ToRow b) => Sql -> b -> m (Maybe a)

queryOneRow :: (DB m, FromRow b, ToRow a) => Sql -> a -> m b

queryOneCol :: (DB m, FromField b, ToRow a) => Sql -> a -> m b

With checks

queryListRowCheck :: (DB m, FromRow b, ToRow a, SqliteExceptionReason e) => Sql -> a -> ([b] -> Either e r) -> m r

queryListColCheck :: (DB m, FromField b, ToRow a, SqliteExceptionReason e) => Sql -> a -> ([b] -> Either e r) -> m r

queryMaybeRowCheck :: (DB m, FromRow b, ToRow a, SqliteExceptionReason e) => Sql -> a -> (Maybe b -> Either e r) -> m r

queryMaybeColCheck :: (DB m, FromField b, ToRow a, SqliteExceptionReason e) => Sql -> a -> (Maybe b -> Either e r) -> m r

queryOneRowCheck :: (DB m, FromRow b, ToRow a, SqliteExceptionReason e) => Sql -> a -> (b -> Either e r) -> m r

queryOneColCheck :: (DB m, FromField b, ToRow a, SqliteExceptionReason e) => Sql -> a -> (b -> Either e r) -> m r

Without parameters

queryListRow_ :: (DB m, FromRow a) => Sql -> m [a]

queryListCol_ :: (DB m, FromField a) => Sql -> m [a]

queryMaybeRow_ :: (DB m, FromRow a) => Sql -> m (Maybe a)

queryMaybeCol_ :: (DB m, FromField a) => Sql -> m (Maybe a)

queryOneRow_ :: (DB m, FromRow a) => Sql -> m a

queryOneCol_ :: (DB m, FromField a) => Sql -> m a

With checks

queryListRowCheck_ :: (DB m, FromRow a, SqliteExceptionReason e) => Sql -> ([a] -> Either e r) -> m r

queryListColCheck_ :: (DB m, FromField a, SqliteExceptionReason e) => Sql -> ([a] -> Either e r) -> m r

queryMaybeRowCheck_ :: (DB m, FromRow a, SqliteExceptionReason e) => Sql -> (Maybe a -> Either e r) -> m r

queryMaybeColCheck_ :: (DB m, FromField a, SqliteExceptionReason e) => Sql -> (Maybe a -> Either e r) -> m r

queryOneRowCheck_ :: (DB m, FromRow a, SqliteExceptionReason e) => Sql -> (a -> Either e r) -> m r

queryOneColCheck_ :: (DB m, FromField a, SqliteExceptionReason e) => Sql -> (a -> Either e r) -> m r

Data version

newtype DataVersion

Constructors

DataVersion Int64

getDataVersion :: Transaction DataVersion

Journal mode

data JournalMode

https://www.sqlite.org/pragma.html#pragma_journal_mode

Constructors

JournalMode'DELETE

JournalMode'TRUNCATE

JournalMode'PERSIST

JournalMode'MEMORY

JournalMode'WAL

JournalMode'OFF

trySetJournalMode :: JournalMode -> Transaction ()

Low-level

withSavepoint :: (DB m, MonadUnliftIO m) => Text -> (m () -> m a) -> m a

Perform an action within a named savepoint. The action is provided a rollback action.

withStatement :: (DB m, MonadUnliftIO m, FromRow a, ToRow b) => Sql -> b -> (m (Maybe a) -> m c) -> m c

Exceptions

data SomeSqliteException

The root exception for all exceptions thrown by this library.

SomeException (from base)
  └── SomeSqliteException
        └── SqliteConnectException
        └── SqliteQueryException

A SomeSqliteException should not be inspected or used for control flow when run in a trusted environment, where the database can be assumed to be uncorrupt. Rather, wherever possible, the user of this library should write code that is guaranteed not to throw exceptions, by checking the necessary preconditions first. If that is not possible, it should be considered a bug in this library.

When actions are run on an untrusted codebase, e.g. one downloaded from a remote server, it is sufficient to catch just one exception type, SomeSqliteException.

Constructors

forall e.Exception e => SomeSqliteException e

data SqliteConnectException

An exception thrown during establishing a connection.

Constructors

SqliteConnectException

Fields

threadId :: ThreadId

name :: String

file :: FilePath

exception :: SQLError

data SqliteQueryException

A SqliteQueryException represents an exception thrown during processing a query, paired with some context that resulted in the exception.

A SqliteQueryException may result from a number of different conditions:

The underlying sqlite library threw an exception.

A postcondition violation of a function like queryMaybeRow, which asserts that the resulting relation will have certain number of rows,

A postcondition violation of a function like queryListRowCheck, which takes a user-defined check as an argument.

Constructors

SqliteQueryException

Fields

threadId :: ThreadId

connection :: String

sql :: Sql

params :: String

exception :: SomeSqliteExceptionReason

The inner exception. It is intentionally not SomeException, so that calling code cannot accidentally throwIO domain-specific exception types, but must instead use a *Check query variant.

class (Show e, Typeable e) => SqliteExceptionReason e

A type that is intended to be used as additional context for a sqlite-related exception.

data SomeSqliteExceptionReason

Constructors

forall e.SqliteExceptionReason e => SomeSqliteExceptionReason e

newtype ExpectedAtMostOneRowException

A query was expected to return exactly one row, but it did not. The exception carries a string representation of the rows that were actually returned.

Constructors

ExpectedAtMostOneRowException

Fields

rows :: String

newtype ExpectedExactlyOneRowException

A query was expected to return exactly one row, but it did not. The exception carries a string representation of the rows that were actually returned.

Constructors

ExpectedExactlyOneRowException

Fields

rows :: String

data SetJournalModeException

Constructors

SetJournalModeException

Fields

currentJournalMode :: JournalMode

couldntSetTo :: JournalMode