Unison.Sqlite.Connection

Connection management

data Connection

A non-thread safe connection to a SQLite database.

Constructors

Connection

Fields

name :: String

file :: FilePath

conn :: Connection

withConnection

Arguments

:: MonadUnliftIO m

=> String
Connection name, for debugging.

-> FilePath
Path to SQLite database file.

-> (Connection -> m a)

-> m a

Perform an action with a connection to a SQLite database.

Note: the connection is created with PRAGMA foreign_keys = ON automatically, to work around the fact that SQLite does not automatically enforce foreign key integrity, because it elected to maintain backwards compatibility with code that was written before the foreign key integrity feature was implemented.

Executing queries

Without results

With parameters

execute :: ToRow a => Connection -> Sql -> a -> IO ()

executeMany :: ToRow a => Connection -> Sql -> [a] -> IO ()

Without parameters

execute_ :: Connection -> Sql -> IO ()

With results

With parameters

queryListRow :: (FromRow b, ToRow a) => Connection -> Sql -> a -> IO [b]

queryListCol :: forall a b. (FromField b, ToRow a) => Connection -> Sql -> a -> IO [b]

queryMaybeRow :: (ToRow a, FromRow b) => Connection -> Sql -> a -> IO (Maybe b)

queryMaybeCol :: forall a b. (ToRow a, FromField b) => Connection -> Sql -> a -> IO (Maybe b)

queryOneRow :: (FromRow b, ToRow a) => Connection -> Sql -> a -> IO b

queryOneCol :: forall a b. (FromField b, ToRow a) => Connection -> Sql -> a -> IO b

With checks

queryListRowCheck :: (FromRow b, ToRow a, SqliteExceptionReason e) => Connection -> Sql -> a -> ([b] -> Either e r) -> IO r

queryListColCheck :: forall a b e r. (FromField b, ToRow a, SqliteExceptionReason e) => Connection -> Sql -> a -> ([b] -> Either e r) -> IO r

queryMaybeRowCheck :: (FromRow b, ToRow a, SqliteExceptionReason e) => Connection -> Sql -> a -> (Maybe b -> Either e r) -> IO r

queryMaybeColCheck :: forall a b e r. (FromField b, ToRow a, SqliteExceptionReason e) => Connection -> Sql -> a -> (Maybe b -> Either e r) -> IO r

queryOneRowCheck :: (FromRow b, ToRow a, SqliteExceptionReason e) => Connection -> Sql -> a -> (b -> Either e r) -> IO r

queryOneColCheck :: forall a b e r. (FromField b, ToRow a, SqliteExceptionReason e) => Connection -> Sql -> a -> (b -> Either e r) -> IO r

Without parameters

queryListRow_ :: FromRow a => Connection -> Sql -> IO [a]

queryListCol_ :: forall a. FromField a => Connection -> Sql -> IO [a]

queryMaybeRow_ :: FromRow a => Connection -> Sql -> IO (Maybe a)

queryMaybeCol_ :: forall a. FromField a => Connection -> Sql -> IO (Maybe a)

queryOneRow_ :: FromRow a => Connection -> Sql -> IO a

queryOneCol_ :: forall a. FromField a => Connection -> Sql -> IO a

With checks

queryListRowCheck_ :: (FromRow a, SqliteExceptionReason e) => Connection -> Sql -> ([a] -> Either e r) -> IO r

queryListColCheck_ :: forall a e r. (FromField a, SqliteExceptionReason e) => Connection -> Sql -> ([a] -> Either e r) -> IO r

queryMaybeRowCheck_ :: (FromRow a, SqliteExceptionReason e) => Connection -> Sql -> (Maybe a -> Either e r) -> IO r

queryMaybeColCheck_ :: forall a e r. (FromField a, SqliteExceptionReason e) => Connection -> Sql -> (Maybe a -> Either e r) -> IO r

queryOneRowCheck_ :: (FromRow a, SqliteExceptionReason e) => Connection -> Sql -> (a -> Either e r) -> IO r

queryOneColCheck_ :: forall a e r. (FromField a, SqliteExceptionReason e) => Connection -> Sql -> (a -> Either e r) -> IO r

Low-level operations

withSavepoint :: Connection -> Text -> (IO () -> IO a) -> IO a

Perform an action within a named savepoint. The action is provided a rollback action.

withStatement :: (FromRow a, ToRow b) => Connection -> Sql -> b -> (IO (Maybe a) -> IO c) -> IO c

Exceptions

newtype ExpectedAtMostOneRowException

A query was expected to return exactly one row, but it did not. The exception carries a string representation of the rows that were actually returned.

Constructors

ExpectedAtMostOneRowException

Fields

rows :: String

newtype ExpectedExactlyOneRowException

A query was expected to return exactly one row, but it did not. The exception carries a string representation of the rows that were actually returned.

Constructors

ExpectedExactlyOneRowException

Fields

rows :: String